V2.8.0功能上线

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/sharereward/ShareRewardAddParam.java
@@ -11,7 +11,6 @@ import java.sql.Timestamp;
 /**
  * @author liufei
  * @date 2019/8/20
- * @description
  */
 @Data
 public class ShareRewardAddParam {
@@ -44,7 +43,7 @@ public class ShareRewardAddParam {
      * 活动优先级
      */
     @NotNull
-    private Short priority;
+    private Integer priority;
     /**
      * 触发条件：1.分享全部商品，2.分享指定商品，3.分享访问量较少商品
      */
